minimal fix for int overflow

+++ commands/mode/Mode.hpp
@@ -36,7 +36,7 @@ class   Mode: public Cmd
         void    unsetKey(Channel*    channel) const;
         void    setTopicRight(Channel*    channel) const;
         void    unsetTopicRight(Channel*    channel) const;
-        void    setUserLimit(unsigned long long int size, Channel*    channel) const;
+        void    setUserLimit(int size, Channel*    channel) const;
         void    unsetUserLimit(Channel*    channel) const;
         void    setUserPrivilege(Client& c, Channel*    channel) const;
         void    unsetUserPrivilege(Client& c, Channel*    channel)  const;
